Integrate functionapp.test.ts into createAzureResource.test.ts
